+++ capstone/src/main/java/com/techelevator/view/Chip.java
@@ -1,12 +1,13 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Chip extends Product{
     public List<Chip> product = new ArrayList<>();
 
-    public Chip(String slotLocation, String name, double price, String category) {
+    public Chip(String slotLocation, String name, BigDecimal price, String category) {
         super(slotLocation, name, price, category);
     }
